initial implementation of transformer
Multi-headed attention on a character sequence model. I'm hoping to
limit this project to an older RTX 2060 with only 6GB VRAM so we have
just 12 layers. The tiny character vocab helps a ton.

Starting this on my default test dataset of Infinite Jest by David
Foster Wallace. It's perfect because, unlike Shakespear who writes in
perfect rhyming iambic pentameters with predictable overall structure
David Foster Wallace breaks absolutely every rule of writing in English.

Fun fact: It's impossible to tell when you overfit to this dataset
because you may get gibberish and then you look for it in the source and
Infinite Jest is gibberish as well.

For example, this:

    ... surrender to is the acknowledged primacy of straight-line
    pursuing this flat and short-sighted idea of personal happiness:

    'The happy pleasure of the person alone, yes?'

    'Except why do you let deLint tie Pemulis and Shaw's shoes to the lines,
    if the lines aren't boundaries?'